Refactor inventory view
